refactor tests setup into plugins and fixtures
